app/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.7
boto3==1.26.0
cryptography==39.0.1
aws-encryption-sdk==2.3.0
